Change: Tests to be placed in separate files
