+++ test/cars/init.yml
@@ -1,0 +1,1 @@
+trucks: true
