+++ test/stars/sol.yml
